triton-server/requirements.txt
# requirements.txt
numpy>=1.21.0
tritonclient[http]>=2.29.0
requests>=2.25.0
opencv-python>=4.5.0
Pillow>=8.0.0
onnx>=1.12.0  # If you want to inspect ONNX models
